[frontendActions][slide-in] Move slide in backgroundable to separate class
